requirements.txt
openai-whisper
# Optional, helpful utilities
tqdm
